requirements.txt
django
django-ckeditor
django-import-export
django-ratelimit
requests
pillow
